refactor: move structs to separate modules
