+++ packages_rs/nextclade/src/gene/mod.rs
@@ -1,4 +1,7 @@
 pub mod cds;
+pub mod cds_segment;
+pub mod frame;
 pub mod gene;
 pub mod genotype;
+pub mod phase;
 pub mod protein;
